requirements.txt
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.3
Pillow==8.3.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-Levenshtein==0.12.2
six==1.16.0
